+++ frontend/src/lib/components/AppShortcuts/AppShortcut.tsx
@@ -1,164 +1,137 @@
-import { useActions, useValues } from 'kea'
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
+import { useActions } from 'kea'
+import React, {
+    ReactNode,
+    cloneElement,
+    forwardRef,
+    isValidElement,
+    useCallback,
+    useEffect,
+    useRef,
+    useState,
+} from 'react'
 
-import { Tooltip } from '@posthog/lemon-ui'
-
-import { sceneLogic } from 'scenes/sceneLogic'
-import { Scene } from 'scenes/sceneTypes'
+import { isMac } from 'lib/utils'
+import { cn } from 'lib/utils/css-classes'
 
 import { KeyboardShortcut } from '~/layout/navigation-3000/components/KeyboardShortcut'
-import { HotKeyOrModifier } from '~/types'
 
-import { appShortcutLogic } from './appShortcutLogic'
+import { AppShortcutType, appShortcutLogic } from './appShortcutLogic'
 
-interface BaseShortcut {
-    /** Array of keys that make up the shortcut (e.g., ['command', 'e']) */
-    keys: HotKeyOrModifier[]
-    /** Description of what this shortcut does */
-    description: string
-    /** Scene this shortcut belongs to */
-    sceneKey?: Scene
-    /** Whether the shortcut is currently enabled */
-    enabled?: boolean
-    /** Whether the shortcut is currently in active state (required for toggle shortcuts) */
-    active?: boolean
-    /** Whether the action palette should close when this action is triggered @default true */
-    closeActionPaletteOnAction?: boolean
-    /** Type of shortcut for icon display */
-    type?: 'action' | 'toggle' | 'link'
-    /** Order for sorting within groups. Lower numbers appear first. -1 = first, 0 = default, 1+ = later */
-    order?: number
+const IS_MAC = isMac()
+
+// Helper function to convert keybind array to KeyboardShortcut props
+function keybindToKeyboardShortcutProps(keybind: string[]): Record<string, boolean> {
+    const platformAgnosticKeybind = keybind.map((key) => (!IS_MAC && key === 'command' ? 'ctrl' : key))
+    return Object.fromEntries(platformAgnosticKeybind.map((key) => [key, true]))
 }
 
-export interface AppShortcut extends BaseShortcut {
-    id: string
-    enabled: boolean
-    action: () => void
-    element?: HTMLElement
-    actionToggle?: (active: boolean) => void
+interface AppShortcutProps extends React.HTMLAttributes<HTMLElement>, Omit<AppShortcutType, 'ref'> {
+    children: ReactNode
+    asChild?: boolean
+    className?: string
+    disabled?: boolean
 }
 
-export interface AppShortcutProps extends BaseShortcut {
-    /** The action to perform when the shortcut is triggered */
-    onAction?: () => void
-    /** Toggle action that receives the current active state - requires active prop to be set */
-    onActionToggle?: (active: boolean) => void
-    /** Optional unique identifier - will auto-generate if not provided */
-    id?: string
-    /** The child element to wrap */
-    children: React.ReactNode
-}
+export const AppShortcut = forwardRef<HTMLElement, AppShortcutProps>(
+    (
+        {
+            children,
+            asChild = false,
+            name,
+            keybind,
+            intent,
+            interaction,
+            scope = 'global',
+            className,
+            disabled = false,
+            ...props
+        },
+        forwardedRef
+    ): JSX.Element => {
+        const internalRef = useRef<HTMLElement>(null)
+        const [isRefReady, setIsRefReady] = useState(false)
+        const { registerAppShortcut, unregisterAppShortcut } = useActions(appShortcutLogic)
 
-export function AppShortcut({
-    keys,
-    description,
-    sceneKey,
-    enabled = true,
-    onAction,
-    onActionToggle,
-    active,
-    id,
-    children,
-    closeActionPaletteOnAction = true,
-    type,
-    order,
-}: AppShortcutProps): JSX.Element {
-    const elementRef = useRef<HTMLElement>(null)
-    const { registerAppShortcut, unregisterAppShortcut } = useActions(appShortcutLogic)
-    const { activeTabId } = useValues(sceneLogic)
-    const { optionKeyHeld } = useValues(appShortcutLogic)
-    const [showTooltip, setShowTooltip] = useState(false)
+        // Use callback ref to track when element is ready
+        const handleRef = useCallback(
+            (node: HTMLElement | null) => {
+                // Handle internal ref
+                ;(internalRef as React.MutableRefObject<HTMLElement | null>).current = node
+                setIsRefReady(!!node)
 
-    // Stable ID that doesn't change on re-renders
-    const shortcutId = useMemo(
-        () => id || `shortcut-${sceneKey}-${keys.join('-')}-${Math.random().toString(36).substring(2, 11)}`,
-        [id, sceneKey, keys]
-    )
-
-    // Stable action callback with toggle logic
-    const stableOnAction = useCallback(() => {
-        if (onActionToggle && active !== undefined) {
-            // Toggle logic: pass current active state to the toggle function
-            onActionToggle(active)
-        } else if (onAction) {
-            // Fallback to traditional onAction
-            onAction()
-        }
-    }, [onAction, onActionToggle, active])
-
-    // Stable shortcut object
-    const shortcut = useMemo(
-        (): AppShortcut => ({
-            id: shortcutId,
-            keys,
-            description,
-            enabled,
-            action: stableOnAction,
-            sceneKey,
-            element: elementRef.current || undefined,
-            closeActionPaletteOnAction,
-            active,
-            actionToggle: onActionToggle,
-            type,
-            order,
-        }),
-        [
-            shortcutId,
-            keys,
-            description,
-            enabled,
-            stableOnAction,
-            sceneKey,
-            closeActionPaletteOnAction,
-            active,
-            onActionToggle,
-            type,
-            order,
-        ]
-    )
-
-    // Register/unregister shortcut
-    useEffect(() => {
-        if (!activeTabId) {
-            return
-        }
-
-        registerAppShortcut(activeTabId, shortcut)
-
-        return () => {
-            unregisterAppShortcut(activeTabId, shortcutId)
-        }
-    }, [activeTabId, shortcut, shortcutId, registerAppShortcut, unregisterAppShortcut])
-
-    return (
-        <div
-            className="contents"
-            onMouseOver={() => {
-                setShowTooltip(true)
-            }}
-            onMouseOut={() => {
-                setShowTooltip(false)
-            }}
-        >
-            <Tooltip
-                title={
-                    showTooltip ? (
-                        <>
-                            {description}{' '}
-                            <KeyboardShortcut
-                                {...Object.fromEntries(keys.map((key) => [key, true]))}
-                                className="text-xs"
-                            />
-                        </>
-                    ) : undefined
+                // Handle forwarded ref
+                if (typeof forwardedRef === 'function') {
+                    forwardedRef(node)
+                } else if (forwardedRef) {
+                    forwardedRef.current = node
                 }
-                visible={(optionKeyHeld || showTooltip) && enabled}
-                closeDelayMs={0}
-            >
-                {React.cloneElement(children as React.ReactElement, {
-                    ref: elementRef,
-                })}
-            </Tooltip>
-        </div>
-    )
-}
+            },
+            [forwardedRef]
+        )
+
+        // Register shortcut when ref is ready
+        useEffect(() => {
+            if (isRefReady && internalRef.current && !disabled) {
+                // Replace 'command' with 'ctrl' when not on Mac
+                const platformAgnosticKeybind = keybind.map((key) => (!IS_MAC && key === 'command' ? 'ctrl' : key))
+                registerAppShortcut({
+                    name,
+                    keybind: platformAgnosticKeybind,
+                    ref: internalRef,
+                    intent,
+                    interaction,
+                    scope,
+                })
+            }
+        }, [isRefReady, name, keybind, intent, interaction, scope, disabled, registerAppShortcut])
+
+        // Clean up on unmount
+        useEffect(() => {
+            return () => {
+                unregisterAppShortcut(name)
+            }
+        }, [name, unregisterAppShortcut])
+
+        const elementProps = {
+            'data-shortcut-name': name,
+            'data-shortcut-keybind': keybind.join('+'),
+            'data-shortcut-intent': intent,
+            'aria-keyshortcuts': keybind.join('+'),
+            ref: handleRef,
+            className: cn(className),
+            ...props,
+        }
+
+        if (asChild && isValidElement(children)) {
+            const childProps = children.props as any
+            let finalTooltip = undefined
+
+            // If the child has a tooltip prop and not disabled, append the keyboard shortcut to it
+            if (childProps.tooltip && !disabled) {
+                finalTooltip = (
+                    <>
+                        {childProps.tooltip}{' '}
+                        <KeyboardShortcut
+                            {...keybindToKeyboardShortcutProps(keybind)}
+                            className="relative text-xs -top-px"
+                        />
+                    </>
+                )
+            } else if (childProps.tooltip) {
+                // If disabled, just use the original tooltip without keyboard shortcut
+                finalTooltip = childProps.tooltip
+            }
+
+            return cloneElement(children as React.ReactElement, {
+                ...children.props,
+                ...elementProps,
+                tooltip: finalTooltip,
+                className: cn(children.props.className, className),
+            })
+        }
+
+        return <div {...(elementProps as React.HTMLAttributes<HTMLDivElement>)}>{children}</div>
+    }
+)
+
+AppShortcut.displayName = 'AppShortcut'
